Refactored the entire project
