+++ Context/taskContext.jsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+
+const TaskContext = createContext();
+
+export default TaskContext;
