+++ constant.js
@@ -1,0 +1,8 @@
+
+
+
+
+const constants = {
+    serverUrl: 'http://localhost:3003',
+};
+export default constants;
