+++ src/Pages/ProfilePage.jsx
@@ -1,0 +1,9 @@
+
+
+const ProfilePage = () => {
+    return (
+        <div> Profile page </div>
+    )
+}
+
+export default ProfilePage
